feat(expressions): Add support for spel functions provided by plugins
